Identified startup problem. Commit before making major changes.
When the app starts up, the MainWindow constructor is called. Since this is the only guaranteed method to be called during startup, I put all of the actions necessary to start the game inside of this constructor. However, starting a new game requires prompting the user for input for the MakeHero method. The user cannot be prompted at this point because the program is still inside of the MainWindow constructor. Thus, the constructor would not have returned and therefore there is not yet any window with which to prompt the user.
Possible solutions:
-- Have the MainWindow constructor still start a new game; however, make it do so by starting a new thread. That new thread would wait until the UI elements are displayed, and then start the game and prompt the user.
-- Make a "Start New Game" button that could be pressed by the user after the window is displayed. This would make the user have to press the button each time the program starts.
-- Find some way to automatically call a StartNewGame method after the MainWindow constructor returns, on the same thread.